requirements.txt
coverage==4.5.4
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
geopandas==0.5.0
Shapely==1.6.4.post2